+++ ui/src/features/goal/components/session-card.tsx
