+++ ui/src/features/goal/components/session-performance.tsx
