+++ ui/src/routes/goals/$id.tsx
@@ -1,7 +1,7 @@
 import { createFileRoute, redirect } from '@tanstack/react-router';
 
 import { fetchMe } from '@/features/auth/api';
-import Plan from '@/features/goals/components/plan';
+import Plan from '@/features/goal/components/plan';
 
 export const Route = createFileRoute('/goals/$id')({
   beforeLoad: async () => {
